# Conflicts:
#	Communication/Data_transferring/TransferData.cpp
#	Communication/Data_transferring/TransferData.h
#	Communication/Data_transferring/TransferData_test.cpp
#	Communication/Socket/Server_function.cpp
#	Communication/Socket/Server_function.h
#	Communication/Socket/Socket_test.cpp
#	communication/communication/communication.vcxproj
#	communication/communication/communication.vcxproj.filters

+++ communication/Validation/Convert.cpp
@@ -12,7 +12,7 @@ vector<uint8_t> Convert::convertCharToUint8(char c)
             res.push_back(0);
         c = c >> 1;
     }
-    for (int i = res.size(); i < 8; i++)
+    for (size_t i = res.size(); i < 8; i++)
         res.push_back(0);
     reverse(res.begin(), res.end());
     return res;
@@ -21,10 +21,10 @@ vector<uint8_t> Convert::convertCharToUint8(char c)
 vector<uint8_t> Convert::convertCharVecToUint8(vector<char> c)
 {
      vector<uint8_t> res,one;
-     int len = c.size();
-     for (int i = 0; i < len; i++) {
+     size_t len = c.size();
+     for (size_t i = 0; i < len; i++) {
          one = convertCharToUint8(c[i]);
-         for (int j = 0; j < 8; j++) {
+         for (size_t j = 0; j < 8; j++) {
              res.push_back(one[j]);
          }
      }
@@ -37,7 +37,7 @@ vector<char> Convert::convertUint8ToChar(vector<uint8_t> c)
     vector<char> res;
     for (auto i = c.begin(); i < c.end();) {
         ch = 0;
-        for (int j = 0; j < 8 && i < c.end(); j++, i++) {
+        for (size_t j = 0; j < 8 && i < c.end(); j++, i++) {
             ch = ch << 1;
             if (*i)
                 ch = ch | 1;
